Refactor update logic into a separate module
